eval: Move encode.c to eval/encode.c
